// firebase/Models/User.ts
// userModel.ts
export interface User {
    about: string;
    created_at: string;
    dateOfBirth: string;
    designation: string;
    easyQuestions: number;
    email: string;
    follower: number;
    following: number;
    gender: string;
    hardQuestions: number;
    id: string;
    image: string;
    isApproved: boolean;
    isContentCreator: boolean;
    is_online: boolean;
    last_active: string;
    lives: number;
    mediumQuestions: number;
    name: string;
    phoneNumber: string;
    place: string;
    profession: string;
    push_token: string;
  }
  